requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-rest-swagger
drf-yasg
django-filter
pytest